added repeated test. Removed obsolete retry tests. Updated soft assertions tests
